Move Resnet34Backbone into image classification app
