+++ ext/image/fl/models/CMakeLists.txt
@@ -4,4 +4,5 @@ target_sources(
   flashlight
   PRIVATE
   ${CMAKE_CURRENT_LIST_DIR}/Resnet.cpp
+  ${CMAKE_CURRENT_LIST_DIR}/Resnet34Backbone.h
 )
